Fold runner postprocess factory into stream
